+++ CHANGELOG.md
@@ -3,7 +3,11 @@
 --------------------------------------------------------------------------------
 
 ## Unreleased
-  
+
+## v1.0.4 - 2024-10-18
+- CGViewBUilder.js: Dashes and periods in sequences are replaced by N's
+- Compare qualifier translation with translated from the sequence
+
 ## v1.0.3 - 2024-10-15
 - CSV maxColumns (default 30). Only look at first maxColumns columns. 
 - Fix contig names for fasta files (use seqID, not name)